Old Code/APIs/Eye Disease/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
Flask==2.1.3
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
timm==0.6.5
torch==1.12.0+cpu
torchaudio==0.12.0+cpu
torchvision==0.13.0+cpu
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
